Renamed the validator modules.
