+++ crates/bld_runner/src/validator/v2.rs
